fix/relations
se agrega la relacion entre customer y predictions

+++ backend/src/main/java/com/team42/churninsight/customer/dto/CriticalCustomerResponse.java
@@ -3,8 +3,10 @@ package com.team42.churninsight.customer.dto;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.team42.churninsight.economic.ValueCustomer;
 
+import java.math.BigDecimal;
+
 public record CriticalCustomerResponse(
         @JsonProperty("customer_id") String customerId,
         @JsonProperty("economic_value") ValueCustomer economicValue,
-        @JsonProperty("priority_score") Double priorityScore
+        @JsonProperty("priority_score") BigDecimal priorityScore
 ) {}
